fix: Source contract  filter
--bug=1065467@tapd-34675357 --user=宋昌昌 【回款记录】新建/编辑回款记录-点击合同名称，弹出选择合同页面-选择合同页面出现作废/归档/不通过/提审/撤销的合同 https://www.tapd.cn/34675357/s/1827445
