+++ backend/crm/src/main/java/cn/cordys/common/service/DataInitService.java
@@ -5,8 +5,8 @@ import cn.cordys.common.util.OnceInterface;
 import cn.cordys.common.util.OnceInterfaceAction;
 import cn.cordys.crm.clue.service.ClueService;
 import cn.cordys.crm.system.domain.Parameter;
+import cn.cordys.crm.system.service.ModuleFieldExtService;
 import cn.cordys.crm.system.service.ModuleFieldService;
-import cn.cordys.crm.system.service.ModuleFormExtService;
 import cn.cordys.crm.system.service.ModuleFormService;
 import cn.cordys.crm.system.service.ModuleService;
 import cn.cordys.mybatis.BaseMapper;
@@ -32,7 +32,7 @@ public class DataInitService {
     @Resource
     private ModuleFormService moduleFormService;
 	@Resource
-	private ModuleFormExtService moduleFormExtService;
+	private ModuleFieldExtService moduleFieldExtService;
     @Resource
     private BaseMapper<Parameter> parameterMapper;
     @Resource
@@ -59,7 +59,7 @@ public class DataInitService {
             initOneTime(moduleFormService::initUpgradeForm, "init.upgrade.form.v1.4.0");
             initOneTime(moduleFormService::initUpgradeForm, "init.upgrade.form.v1.5.0");
             initOneTime(moduleFormService::initExtFieldsByVer, "1.5.0", "init.ext.fields.v1.5.0");
-			initOneTime(moduleFormExtService::setDefaultOptionSource, "set.default.option.source");
+			initOneTime(moduleFieldExtService::setDefaultOptionSource, "set.default.option.source");
         } finally {
             lock.unlock();
         }
