+++ backend/crm/src/main/java/cn/cordys/crm/system/controller/ModuleFieldController.java
@@ -157,6 +157,7 @@ public class ModuleFieldController {
     @PostMapping("/source/contract")
     @Operation(summary = "分页获取合同")
     public Pager<List<ContractListResponse>> sourceContractPage(@Valid @RequestBody ContractPageRequest request) {
+		request.setCombineSearch(contractService.setDefaultSourceCombine());
         request.setCombineSearch(request.getCombineSearch().convert());
         DeptDataPermissionDTO deptDataPermission = dataScopeService.getDeptDataPermission(SessionUtils.getUserId(), OrganizationContext.getOrganizationId(), InternalUserView.ALL.name(),
                 PermissionConstants.CONTRACT_READ);
